refactor(copilot): extract completion billing helpers
Move completion billing and mirror logic into a shared lib module used by copilot routes.\n\nCo-authored-by: Codex <codex@openai.com>
